Orientation turnRate: promote to double + exact reverse calc
Turn rate is JUnits per CLIENT tick, so 4 over 8 game ticks only sums
to 4*240=960 JUnits (~169deg) -- nowhere near a full 2048 turn. Hitting
2048 in 8 game ticks needs ~8.533, which integer turn rate couldn't
express.

* OrientationKeyFrame.turnRate: int -> double. TURN_RATE constant
  stays int so Movement's default keeps compiling unchanged.
* Spinner model: double-typed, step 0.1, "0.###" formatter so the
  converted value shows up to 3 decimals.
* calculateOrientationTurnRate now returns the exact double rounded
  to 0.001 instead of evaluating floor/ceil and picking the closer
  integer.
* Debug log: turnRate=%d -> %.3f for orientation activations.

+++ src/main/java/com/creatorskit/swing/timesheet/TimeSheetPanel.java
@@ -1251,7 +1251,7 @@ public class TimeSheetPanel extends JPanel
         selectedCharacter.setVisible(true, clientThread);
     }
 
-    public void initializeOrientationKeyFrame(Character character, OrientationHotkeyMode hotkeyMode, LocalPoint localPoint, double tick, int start, int end, OrientationGoal og, int turnRate)
+    public void initializeOrientationKeyFrame(Character character, OrientationHotkeyMode hotkeyMode, LocalPoint localPoint, double tick, int start, int end, OrientationGoal og, double turnRate)
     {
         CKObject ckObject = character.getCkObject();
         if (ckObject == null)
